Catchup

+++ index.js
@@ -103,5 +103,5 @@ bot.on("message", msg => {
     msg.reply("there was an error trying to execute that command!");
   }
 
-  msg.delete();
+  if(msg.channel.type != "dm") msg.delete();
 });
